requirements.txt
faker>=25.0.0
kafka-python>=2.0.2
pandas>=2.2.0
psycopg2-binary>=2.9.9
pyspark>=3.5.1
flask>=3.0.0
streamlit>=1.33.0
python-dotenv>=1.0.1
